requirements.txt
fastapi==0.115.11
uvicorn==0.34.0
sqlalchemy==2.0.38
psycopg2-binary==2.9.10
python-dotenv==1.0.1
pytest==8.3.5
httpx==0.28.1
jinja2==3.1.5
python-multipart==0.0.20
pytest-mock==3.14.0
pytest-cov==5.0.0
beautifulsoup4==4.12.3
